backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.6.1
chromadb==0.4.18
langchain==0.0.350
langchain-openai>=0.0.2
PyPDF2==3.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
numpy==1.24.3
tiktoken==0.5.2
